signed-ness errors

+++ vamp-plugins/dsp/TonalEstimator.h
@@ -32,7 +32,7 @@ public:
 	
 	void printDebug()
 	{
-		for (int i = 0; i < size(); i++)
+		for (size_t i = 0; i < size(); i++)
 		{
 			std::cout <<  (*this)[i] << ";";
 		}
@@ -68,7 +68,7 @@ public:
 
 	void printDebug()
 	{
-		for (int i = 0; i < size(); i++)
+		for (size_t i = 0; i < size(); i++)
 		{
 			std::cout <<  (*this)[i] << ";";
 		}
